requirements.txt
Django==2.2.4
django-discover-runner==1.0
django-role-permissions==2.2.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
